requirements.txt
rdkit-pypi
numpy==1.26.4
bokeh
umap-learn
pca
scikit-image
pandas
